+++ src/object/Edge.java
@@ -3,15 +3,15 @@ public class Edge {
 
 	private String fromnode;
 	private String tonode;
-	private double weight;
+	private int weight;
 	private int fromtime;
 	private int totime;
 
-	public double getWeight() {
+	public int getWeight() {
 		return weight;
 	}
 
-	public void setWeight(double weight) {
+	public void setWeight(int weight) {
 		this.weight = weight;
 	}
